+++ EBanking.Services/IAccountService.cs
@@ -5,7 +5,7 @@ namespace EBanking.Services
     public interface IAccountService
     {
         List<AccountModel> GetAccountsForUser(int userId);
-        void UpdateBalance(double newBalance, string accountNumber);
+        void UpdateBalance(decimal newBalance, string accountNumber);
         bool IsValidAccount(string accountNumber);
         Task<AccountModel> GetAccountByAccountNumber(string accountNumber);
     }
